LayoutTests/svg/animations/svgnumberlist-animation-2-expected.txt

SVG 1.1 dynamic animation tests

ABCD
Test 'by' animation of SVGNumberList on 'rotate' attribute of text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text.rotate.animVal.getItem(0).value is 0
PASS text.rotate.animVal.getItem(1).value is 45
PASS text.rotate.animVal.getItem(2).value is 90
FAIL text.rotate.animVal.getItem(3).value should be 135. Was 153.
PASS text.rotate.animVal.getItem(0).value is 22.5
PASS text.rotate.animVal.getItem(1).value is 67.5
PASS text.rotate.animVal.getItem(2).value is 112.5
PASS text.rotate.animVal.getItem(3).value is 157.5
PASS text.rotate.animVal.getItem(0).value is 45
PASS text.rotate.animVal.getItem(1).value is 90
PASS text.rotate.animVal.getItem(2).value is 135
PASS text.rotate.animVal.getItem(3).value is 180
PASS text.rotate.animVal.getItem(0).value is 0
PASS text.rotate.animVal.getItem(1).value is 45
PASS text.rotate.animVal.getItem(2).value is 90
FAIL text.rotate.animVal.getItem(3).value should be 135. Was 153.
PASS successfullyParsed is true

TEST COMPLETE
